+++ packages/atproto/lib/atproto.dart
@@ -66,6 +66,7 @@ export 'package:atproto/src/services/entities/service_auth_token.dart';
 export 'package:atproto/src/services/entities/account_status.dart';
 export 'package:atproto/src/services/entities/missing_blobs.dart';
 export 'package:atproto/src/services/entities/record_blob.dart';
+export 'package:atproto/src/services/entities/plc_operation.dart';
 export 'package:atproto/src/services/identity_service.dart';
 export 'package:atproto/src/services/constants/moderation_reason_type.dart';
 export 'package:atproto/src/services/moderation_service.dart';